refactor: #21 기존 ExceptionCode 객체 명칭을 ErrorCode로 변경
