+++ src/main/java/site/dogether/challengegroup/controller/response/ChallengeGroupErrorCode.java
@@ -2,15 +2,16 @@ package site.dogether.challengegroup.controller.response;
 
 import lombok.Getter;
 import lombok.RequiredArgsConstructor;
-import site.dogether.common.exception.ExceptionCode;
+import site.dogether.common.exception.ErrorCode;
 
 @Getter
 @RequiredArgsConstructor
-public enum ChallengeGroupExceptionCode implements ExceptionCode {
+public enum ChallengeGroupErrorCode implements ErrorCode {
 
     MEMBER_NOT_IN_CHALLENGE_GROUP("CGF-0001"),
     NOT_ENOUGH_CHALLENGE_GROUP_MEMBERS("CGF-0002"),
     NOT_RUNNING_CHALLENGE_GROUP("CGF-0003"),
+    INVALID_CHALLENGE_GROUP_EXCEPTION("CGF-0004");
     ;
 
     private final String value;
