+++ src/main/java/site/dogether/dailytodo/controller/DailyTodoExceptionHandler.java
@@ -9,7 +9,7 @@ import org.springframework.web.bind.annotation.RestControllerAdvice;
 import site.dogether.common.controller.response.ApiResponse;
 import site.dogether.dailytodo.exception.*;
 
-import static site.dogether.dailytodo.controller.response.DailyTodoExceptionCode.*;
+import static site.dogether.dailytodo.controller.response.DailyTodoErrorCode.*;
 
 @Slf4j
 @Order(Ordered.HIGHEST_PRECEDENCE)
