+++ src/main/java/site/dogether/dailytodo/controller/response/DailyTodoErrorCode.java
@@ -2,11 +2,11 @@ package site.dogether.dailytodo.controller.response;
 
 import lombok.Getter;
 import lombok.RequiredArgsConstructor;
-import site.dogether.common.exception.ExceptionCode;
+import site.dogether.common.exception.ErrorCode;
 
 @Getter
 @RequiredArgsConstructor
-public enum DailyTodoExceptionCode implements ExceptionCode {
+public enum DailyTodoErrorCode implements ErrorCode {
 
     INVALID_DAILY_TODO("DTF-0001"),
     UNREVIEWED_DAILY_TODO_EXIST("DTF-0002"),
