+++ src/main/java/site/dogether/notification/controller/NotificationExceptionHandler.java
@@ -9,7 +9,7 @@ import org.springframework.web.bind.annotation.RestControllerAdvice;
 import site.dogether.common.controller.response.ApiResponse;
 import site.dogether.notification.exception.InvalidNotificationTokenException;
 
-import static site.dogether.notification.controller.response.NotificationExceptionCode.INVALID_NOTIFICATION_TOKEN;
+import static site.dogether.notification.controller.response.NotificationErrorCode.INVALID_NOTIFICATION_TOKEN;
 
 @Slf4j
 @Order(Ordered.HIGHEST_PRECEDENCE)
