+++ src/main/java/site/dogether/notification/controller/response/NotificationErrorCode.java
@@ -2,11 +2,11 @@ package site.dogether.notification.controller.response;
 
 import lombok.Getter;
 import lombok.RequiredArgsConstructor;
-import site.dogether.common.exception.ExceptionCode;
+import site.dogether.common.exception.ErrorCode;
 
 @Getter
 @RequiredArgsConstructor
-public enum NotificationExceptionCode implements ExceptionCode {
+public enum NotificationErrorCode implements ErrorCode {
 
     INVALID_NOTIFICATION_TOKEN("NTF-0001")
     ;
